+++ Changelog.md
@@ -24,6 +24,7 @@ Bugfixes:
  * SMTChecker: Fix display error for negative integers that are one more than powers of two.
  * SMTChecker: Improved readability for large integers that are powers of two or almost powers of two in error messages.
  * SMTChecker: Fix internal error when a public library function is called internally.
+ * SMTChecker: Fix internal error on multiple wrong SMTChecker natspec entries.
 
 
 ### 0.8.17 (2022-09-08)